Make it safe to call PartialRegistration.delete
Immediately after deploying https://github.com/code-dot-org/code-dot-org/pull/35915 we started seeing errors from calls to PartialRegistration.delete when a partial registration was not in progress: https://app.honeybadger.io/projects/3240/faults/66466286

This adds a guard for this case and a test covering the offending behavior.

+++ dashboard/app/models/concerns/partial_registration.rb
@@ -35,6 +35,7 @@ module PartialRegistration
   end
 
   def self.delete(session)
+    return unless in_progress? session
     CDO.shared_cache.delete(session[SESSION_KEY])
     session.delete(SESSION_KEY)
   end
